Refactor: Extract saveAndProcessPage tests from action_creators.spec.js
Closes #211
